refactor: extract class for test
